Embed server in service;
