+++ .gitignore
@@ -7,3 +7,4 @@ app/app
 app/cat-signal-app
 .DS_Store
 status-colors.txt
+static.zip
